+++ src/libs/shared/components/menu/menu.component.html
@@ -1,30 +1,18 @@
 <ion-list class="menu">
-  <button *ngIf="!isLoggedIn"
-          menuClose
-          ion-item
-          no-lines
-          (click)="login()">
-    <ion-icon item-start
-              name="ios-log-in"></ion-icon>
+  <button *ngIf="!isLoggedIn" menuClose ion-item no-lines (click)="login()">
+    <ion-icon item-start name="ios-log-in"></ion-icon>
     Login / Cadastre-se
   </button>
   <ng-container *ngFor="let menu of menus">
     <ion-list-header>{{menu.name}}</ion-list-header>
-    <button ion-item
-            *ngFor="let page of menu.items;let last = last"
-            [ngClass]="{ last: last, accessDenied: !isLoggedIn && page.secure }"
-            (click)="openPage(page, !isLoggedIn && page.secure )">
-      <ion-icon item-start
-                name="md-{{page.icon}}"></ion-icon>
+    <button ion-item id="{{menu.name}}" *ngFor="let page of menu.items;let last = last" [ngClass]="{ last: last, accessDenied: !isLoggedIn && page.secure }"
+      (click)="openPage(page, !isLoggedIn && page.secure )">
+      <ion-icon item-start name="md-{{page.icon}}"></ion-icon>
       {{page.title}}
     </button>
   </ng-container>
-  <button *ngIf="isLoggedIn"
-          menuClose
-          ion-item
-          (click)="logout()">
-    <ion-icon item-start
-              name="log-out"></ion-icon>
+  <button *ngIf="isLoggedIn" menuClose ion-item (click)="logout()">
+    <ion-icon item-start name="log-out"></ion-icon>
     Sair
   </button>
 </ion-list>
